Changes the Ubuntu 14.04 dependancy libboost-all-dev to libboost-dev
This fixes an issue when following the insruction to install
dependancies for Ubuntu 14.04, this issue did not effect version
16.04.  However the fix for 14.04 works just as well for 16.04,
outlined at this ask Ubuntu thread;
http://askubuntu.com/questions/670924/how-to-install-boost-on-ubuntu-14-04

The issue it fixes has an example here http://pastebin.com/AAgvYSbQ
blocking the install of libboost1.54 development files.

This fix was tested on 16.04 and 14.04 both of which installed
the depdancies need for the nheqminer.